+++ src/imu.h
@@ -1,7 +1,7 @@
 #ifndef IMU_H
 #define IMU_H
 
-#include <dlib/matrix.h>
+#include "dog_status.h"
 
 class Imu
 {
@@ -10,14 +10,14 @@ public:
     Imu();
     int bus_fd;
 
-    dlib::vector<double> gyroOffset {0.0, 0.0, 0.0};
-    dlib::vector<double> acceleratorOffset {0.0, 0.0, 0.0};
+    Vec3 gyroOffset {};
+    Vec3 acceleratorOffset {};
 
     double accelerometerSensitivityPerBit {16384.0};
     double gyroSensitivityPerBit{131.0};
 
-    dlib::vector<double> readGyro();
-    dlib::vector<double> readAccelerometer();
+    Vec3 readGyro();
+    Vec3 readAccelerometer();
     double readTemperature();
     bool isOpen() const;